The merge config has been refactored
